+++ modules/actors/actor-sheet-ffg.js
@@ -665,9 +665,6 @@ export class ActorSheetFFG extends foundry.appv1.sheets.ActorSheet {
         type: "Boolean",
         default: true,
       });
-      html.find(".source-control").click(async (ev) => {
-        await this._handleSourceControl(ev);
-      });
     }
 
     if (this.actor.type !== "homestead") {
@@ -678,6 +675,14 @@ export class ActorSheetFFG extends foundry.appv1.sheets.ActorSheet {
         default: false,
       });
     }
+
+    // activate source and tag controls for actors
+    html.find(".source-control").click(async (ev) => {
+      await this._handleSourceControl(ev);
+    });
+    html.find(".tag-control").click(async (ev) => {
+      await this._handleTagControl(ev);
+    });
 
     html.find(".medical").click(async (ev) => {
       const item = await $(ev.currentTarget);
@@ -2728,6 +2733,52 @@ export class ActorSheetFFG extends foundry.appv1.sheets.ActorSheet {
       const sources = foundry.utils.deepClone(this.object.system.metadata.sources);
       sources.splice(sourceIndex, 1);
       await this.object.update({"system.metadata.sources": sources});
+    }
+    this.render(true);
+  }
+
+  /**
+   * Handle adding a tag to actors
+   * @param event
+   * @returns {Promise<void>}
+   * @private
+   */
+  async _handleTagControl(event) {
+    event.preventDefault();
+    event.stopPropagation();
+    const action = $(event.currentTarget).data("action");
+    const tagIndex = $(event.currentTarget).data("index");
+    if (action === "add") {
+      const addTag = new Dialog({
+        title: game.i18n.localize("SWFFG.Meta.Tags.AddTag.Title"),
+        content: `
+          <p>${game.i18n.localize("SWFFG.Meta.Tags.AddTag.Tag")} :</p>
+          <input type="text" id="tag" name="tag" value="" autofocus>
+        `,
+        buttons: {
+          submit: {
+            icon: '<i class="fas fa-check"></i>',
+            label: game.i18n.localize("SWFFG.Meta.Tags.AddTag.Submit"),
+            callback: async (obj, event) => {
+              const jObj = $(obj);
+              const tag = jObj.find("#tag").val();
+              const updatedTags = this.object.system.metadata.tags || [];
+              updatedTags.push(tag);
+              await this.object.update({"system.metadata.tags": updatedTags});
+            }
+          },
+          cancel: {
+            icon: '<i class="fas fa-x"></i>',
+            label: game.i18n.localize("SWFFG.Meta.Tags.AddTag.Cancel"),
+          },
+        },
+        default: "submit",
+      });
+      addTag.render(true, {focus: true, classes: ["app", "window-app", "dialog", "themed", "theme-light", "starwarsffg-dialog"]});
+    } else if (action === "remove") {
+      const tags = foundry.utils.deepClone(this.object.system.metadata.tags);
+      tags.splice(tagIndex, 1);
+      await this.object.update({"system.metadata.tags": tags});
     }
     this.render(true);
   }